Consolidate all files in one place.
MinkExtesion is not suite aware so only knows the location of one files folder
per profile. I could have done a PR to fix this in MinkExtension but instead
I've decided that the easiest course of action is to use a convention of
files/<test_suite>  location and symlink back to appropriate sections as needed.

TODO: maybe as a good deed I should consider doing a PR to make MinkExtension
suite aware.